test262: Port expectations file from v4's test262.py
Useful for marking tests as being expected to fail (due to, e.g,
functionality that is not yet implemented or expectations in the test
that cannot actually be met, for whatever reason)

The way this works is that a file called "TestExpectations" is searched
for in the root directory (--test). The contents of this file are
searched if --with-test-expectations is passed (if it is found), with
the following syntax rules:

    # a comment
    some_test_to_skip_here
    some_test_we_expect_to_fail_here failing

These results are respected when processing results.

Furthermore, if --update-expectations is passed, tests that were
failing before (but now pass) will be removed from TestExpectations for
future runs.